+++ index.html
@@ -1,18 +1,29 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
-
 <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite App</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="style.css">
+    <title>Document</title>
 </head>
-
 <body>
+    <div class="header">
+        
+    </div>
+    <section class="first">
+        <div class="wrap">
+            <h1>Добро пожаловать, Alex Adams!</h1>
+            <p class="email">alexadams@google.com</p>
+            <div class="cards">
+                <h2>Мои кошельки</h2>
+                <div class="reload">
+                    
+                </div>
+                
+            </div>
+        </div>
+    </section>
 
-    
-
-    <script type="module" src="/main.js"></script>
+    <script type="module" src="./main.js"></script>
 </body>
-
 </html>
