--- modules/cards_reload/style.scss
@@ -1,99 +1,0 @@
-*{
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-}
-img{
-    max-width: 100%;
-}
-.wrap{
-    width: 1200px;
-    margin: 30px auto;
-}
-.first {
-    h1{
-        font-size: 48px;
-        margin-bottom: 1%;
-        line-height: 59.42px;
-    }
-    .email{
-        cursor: pointer;
-        font-size: 18px;
-        color: #0047FF;
-        font-weight: 700;
-        line-height: 22.28px;
-    }
-    .cards{
-        margin-block: 70px;
-        h2{
-            font-weight: 500;
-            font-size: 24px;
-            line-height: 29.11px;
-        }
-    }
-    .reload{
-        margin-top: 15px;
-        display: grid;
-        grid-template-columns: repeat(4,1fr);
-        .card_visa{
-            position: relative;
-            background: linear-gradient(84.37deg, #D7816A 2.27%, #BD4F6C 92.26%);
-            width: 282px;
-            height: 160px;
-            border-radius: 15px;
-            h3{
-                font-size: 24px;
-                font-weight: 500;
-                line-height: 29px;
-                color: white;
-                margin: 10px 20px;
-            }
-            p{
-                margin: 0 20px;
-                position: absolute;
-                bottom: 10px;
-                color: white;
-            }
-        }
-    }
-
-}
-header{
-    img {
-        width: 20%;
-        height: 20%;
-    }
-    
-    
-    a {
-        color: black;
-    }
-    
-    .box {
-        display: flex;
-        justify-content: space-around;
-        width: 100%;
-        height: 70px;
-        background-color: #F0F0F0;
-        align-items: center;
-       
-    }
-    
-    
-    .btn {
-        color: red;
-        width: 50%;
-        height: 50%;
-        border: none;
-        cursor: pointer;
-    }
-    
-    .box_l {
-        display: flex;
-        gap: 30px;
-    }
-    .box_r {
-        display: flex;
-        gap: 20px;
-    }
-}
