+++ style.css
@@ -1,96 +1,99 @@
-:root {
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-
-  color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
-
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-a {
-  font-weight: 500;
-  color: #646cff;
-  text-decoration: inherit;
-}
-a:hover {
-  color: #535bf2;
-}
-
-body {
+*{
+  padding: 0;
   margin: 0;
-  display: flex;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
+  box-sizing: border-box;
 }
-
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
+img{
+  max-width: 100%;
 }
-
-#app {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
+.wrap{
+  width: 1200px;
+  margin: 30px auto;
 }
-
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.vanilla:hover {
-  filter: drop-shadow(0 0 2em #f7df1eaa);
-}
-
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
-}
-
-button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #1a1a1a;
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-button:hover {
-  border-color: #646cff;
-}
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
-
-@media (prefers-color-scheme: light) {
-  :root {
-    color: #213547;
-    background-color: #ffffff;
+.first {
+  h1{
+      font-size: 48px;
+      margin-bottom: 1%;
+      line-height: 59.42px;
   }
-  a:hover {
-    color: #747bff;
+  .email{
+      cursor: pointer;
+      font-size: 18px;
+      color: #0047FF;
+      font-weight: 700;
+      line-height: 22.28px;
   }
-  button {
-    background-color: #f9f9f9;
+  .cards{
+      margin-block: 70px;
+      h2{
+          font-weight: 500;
+          font-size: 24px;
+          line-height: 29.11px;
+      }
+  }
+  .reload{
+      margin-top: 15px;
+      display: grid;
+      grid-template-columns: repeat(4,1fr);
+      .card_visa{
+          position: relative;
+          background: linear-gradient(84.37deg, #D7816A 2.27%, #BD4F6C 92.26%);
+          width: 282px;
+          height: 160px;
+          border-radius: 15px;
+          h3{
+              font-size: 24px;
+              font-weight: 500;
+              line-height: 29px;
+              color: white;
+              margin: 10px 20px;
+          }
+          p{
+              margin: 0 20px;
+              position: absolute;
+              bottom: 10px;
+              color: white;
+          }
+      }
+  }
+
+}
+header{
+  img {
+      width: 20%;
+      height: 20%;
+  }
+  
+  
+  a {
+      color: black;
+  }
+  
+  .box {
+      display: flex;
+      justify-content: space-around;
+      width: 100%;
+      height: 70px;
+      background-color: #F0F0F0;
+      align-items: center;
+     
+  }
+  
+  
+  .btn {
+      color: red;
+      width: 50%;
+      height: 50%;
+      border: none;
+      cursor: pointer;
+  }
+  
+  .box_l {
+      display: flex;
+      gap: 30px;
+  }
+  .box_r {
+      display: flex;
+      gap: 20px;
   }
 }
